+++ src/components/NProgress/index.ts
@@ -1,0 +1,7 @@
+export { nprogress, NProgressBar } from "./NProgress";
+export type {
+    NProgressBarProps,
+    NProgressController,
+    NProgressListener,
+    NProgressState,
+} from "./NProgress";